refactor(ui-react): relocate StoryDecorator and repoint Storybook specifiers
Step 5 of the DLS-1022 fold.

StoryDecorator.tsx moves from the doomed viz lib into libs/ui-react/.storybook/.
It stays rather than being dropped because it adds `p-32` padding that the
global withStorybookProviders decorator does not -- removing it would produce
eight gratuitous Chromatic diffs on stories that are otherwise unchanged. Its
own ThemeProvider import now uses a relative path, matching the sibling
Decorator.tsx convention rather than the package specifier.

All 11 .storybook/* specifiers in the moved tree are recomputed from each file's
real location rather than nudged by a fixed number of levels: the 8
StoryDecorator imports sat at two different depths, and the 3 DocTable imports
in MDX previously crossed the lib boundary via '../../ui-react/.storybook/...',
which no longer exists.

main.ts drops the two cross-lib globs. No widening is needed -- the existing
'../src/lib/**/*.mdx' and '../src/**/*.stories' globs already reach the
relocated tree, and .storybook/** is already a Tailwind content root.
