+++ libs/ui-react/.storybook/main.ts
@@ -8,8 +8,6 @@ const config: StorybookConfig = {
   stories: [
     '../src/lib/**/*.mdx',
     '../src/**/*.stories.@(js|jsx|ts|tsx|mdx)',
-    '../../ui-react-visualization/src/**/*.mdx',
-    '../../ui-react-visualization/src/**/*.stories.@(js|jsx|ts|tsx|mdx)',
     './docs/**/*.@(mdx)',
     './docs/**/*.stories.@(js|jsx|ts|tsx|mdx)',
   ],
