+++ libs/ui-react/src/lib/Components/visualization/Axis/YAxis/YAxis.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react-vite';
 
-import { StoryDecorator } from '../../../../../.storybook/StoryDecorator';
+import { StoryDecorator } from '../../../../../../.storybook/StoryDecorator';
 import { LineChart } from '../../LineChart';
 import {
   CHART_HEIGHT,
