+++ libs/ui-react/src/lib/Components/visualization/DonutChart/__stories__/DonutChart.mdx
@@ -1,5 +1,5 @@
 import { Meta, Canvas, Controls } from '@storybook/addon-docs/blocks';
-import { DocTable } from '../../../../../../ui-react/.storybook/components';
+import { DocTable } from '../../../../../../.storybook/components';
 import * as DonutChartStories from './DonutChart.stories';
 
 <Meta of={DonutChartStories} />
