+++ libs/ui-react/src/lib/Components/visualization/Legend/__stories__/Legend.mdx
@@ -1,5 +1,5 @@
 import { Meta, Canvas, Controls } from '@storybook/addon-docs/blocks';
-import { DocTable } from '../../../../../../ui-react/.storybook/components';
+import { DocTable } from '../../../../../../.storybook/components';
 import * as LegendStories from './Legend.stories';
 
 <Meta of={LegendStories} />
