+++ libs/ui-react/src/lib/Components/visualization/LineChart/__stories__/LineChart.mdx
@@ -1,5 +1,5 @@
 import { Meta, Canvas, Controls } from '@storybook/addon-docs/blocks';
-import { DocTable } from '../../../../../../ui-react/.storybook/components';
+import { DocTable } from '../../../../../../.storybook/components';
 import * as LineChartStories from './LineChart.stories';
 
 <Meta of={LineChartStories} />
